requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.4.0